Added sweetalert to donate simulation

+++ resources/js/components/admin/guests/GuestsIndex.jsx
@@ -1,11 +1,10 @@
 import React, { useEffect, useState } from 'react';
 import axios from 'axios';
 import { Link } from 'react-router-dom';
-
+import Swal from 'sweetalert2';
 
 function AdminComponent() {
   const [guests, setGuests] = useState([]);
-
 
   useEffect(() => {
     fetchGuests();
@@ -20,15 +19,41 @@ function AdminComponent() {
     }
   };
 
-
   const handleDeleteGuest = async (guestId) => {
-    try {
-      // Skicka ett DELETE-anrop till API för att ta bort gästen med det angivna ID:et
-      await axios.delete(`/api/guests/${guestId}`);
-      // Uppdatera gästlistan efter borttagningen
-      fetchGuests();
-    } catch (error) {
-      console.error('Error deleting guest:', error);
+    // Visa en SweetAlert för bekräftelse
+    const result = await Swal.fire({
+      title: 'Är du säker?',
+      text: 'Gästen kommer tas bort permanent!',
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Ja, ta bort!',
+      cancelButtonText: 'Avbryt',
+    });
+
+    // Om användaren bekräftar borttagningen
+    if (result.isConfirmed) {
+      try {
+        // Skicka ett DELETE-anrop till API för att ta bort gästen med det angivna ID:et
+        await axios.delete(`/api/guests/${guestId}`);
+        // Uppdatera gästlistan efter borttagningen
+        fetchGuests();
+        // Visa en SweetAlert för att informera om att borttagningen är klar
+        Swal.fire({
+          icon: 'success',
+          title: 'Borttagning klar!',
+          text: 'Gästen har tagits bort.',
+        });
+      } catch (error) {
+        console.error('Error deleting guest:', error);
+        // Visa en SweetAlert för felmeddelande
+        Swal.fire({
+          icon: 'error',
+          title: 'Något gick fel!',
+          text: 'Det uppstod ett fel vid borttagning av gästen.',
+        });
+      }
     }
   };
 
@@ -41,8 +66,7 @@ function AdminComponent() {
             <th className="border border-gray-300 py-2 px-4">ID</th>
             <th className="border border-gray-300 py-2 px-4">Namn</th>
             <th className="border border-gray-300 py-2 px-4">Namnbricka</th>
-            <th className="border border-gray-300 py-2 px-4">Inbjudan ID</th>
-            <th className="border border-gray-300 py-2 px-4">Allergier</th>
+            <th className="border border-gray-300 py-2 px-4">Inbjudnigskod</th>
             <th className="border border-gray-300 py-2 px-4">Åtgärder</th>
           </tr>
         </thead>
@@ -52,10 +76,9 @@ function AdminComponent() {
               <td className="border border-gray-300 py-2 px-4">{guest.id}</td>
               <td className="border border-gray-300 py-2 px-4">{guest.name}</td>
               <td className="border border-gray-300 py-2 px-4">{guest.name_tag}</td>
-              <td className="border border-gray-300 py-2 px-4">{guest.invitation_id}</td>
-              <td className="border border-gray-300 py-2 px-4">{guest.allergies}</td>
+              <td className="border border-gray-300 py-2 px-4">{guest.invitation ? guest.invitation.code : 'N/A'}</td>
               <td className="border border-gray-300 py-2 px-4">
-              <Link to={`/admin/guests/${guest.id}`} className="text-blue-500 underline">
+                <Link to={`/admin/guests/${guest.id}`} className="text-blue-500 underline">
                   Visa detaljer
                 </Link>
                 <button
